// frontend/src/client/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: `${string}://${string}` | (string & {});
};

export type Errors = {
    errors?: Array<ErrorMessage>;
};

export type ErrorMessage = {
    code: string;
    message: string;
};

export type Metadata = {
    total: number;
};

export type Category = {
    id: number;
    name: string;
    description?: string;
    emsIds: Array<string>;
};

export type CategoriesResponse = {
    data: Array<Category>;
    metadata: Metadata;
};

export type EmsCategory = {
    id: string;
    name: string;
};

export type EmsCategoriesResponse = {
    data: Array<EmsCategory>;
    metadata: Metadata;
};

export type EmsThemeTranslation = {
    language: string;
    value: string;
    description?: string;
};

export type EmsTheme = {
    code: string;
    translations: Array<EmsThemeTranslation>;
    emsIds: Array<string>;
    datasetsCount: number;
};

export type EmsThemesResponse = {
    data: Array<EmsTheme>;
    metadata: Metadata;
};

export type GetCoreCategoriesData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/v1/core/categories';
};

export type GetCoreCategoriesErrors = {
    /**
     * Error response
     */
    500: Errors;
};

export type GetCoreCategoriesError = GetCoreCategoriesErrors[keyof GetCoreCategoriesErrors];

export type GetCoreCategoriesResponses = {
    /**
     * OK
     */
    200: CategoriesResponse;
};

export type GetCoreCategoriesResponse = GetCoreCategoriesResponses[keyof GetCoreCategoriesResponses];

export type GetCoreEmsCategoriesData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/v1/core/ems-categories';
};

export type GetCoreEmsCategoriesErrors = {
    /**
     * Error response
     */
    500: Errors;
};

export type GetCoreEmsCategoriesError = GetCoreEmsCategoriesErrors[keyof GetCoreEmsCategoriesErrors];

export type GetCoreEmsCategoriesResponses = {
    /**
     * OK
     */
    200: EmsCategoriesResponse;
};

export type GetCoreEmsCategoriesResponse = GetCoreEmsCategoriesResponses[keyof GetCoreEmsCategoriesResponses];

export type GetCoreEmsThemesData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/v1/core/ems-themes';
};

export type GetCoreEmsThemesErrors = {
    /**
     * Error response
     */
    500: Errors;
};

export type GetCoreEmsThemesError = GetCoreEmsThemesErrors[keyof GetCoreEmsThemesErrors];

export type GetCoreEmsThemesResponses = {
    /**
     * OK
     */
    200: EmsThemesResponse;
};

export type GetCoreEmsThemesResponse = GetCoreEmsThemesResponses[keyof GetCoreEmsThemesResponses];